Modernize codebase with centralized configuration and comprehensive lint fixes
## Configuration Management
- Add user_inputs.py with dataclass-based configuration management
- Replace CLI arguments with environment variable configuration
- Add test mode support with TESTING=1 environment variable
- Implement backwards compatibility for GITHUB_TOKEN variables

## Type Safety & Code Quality
- Resolve all 20 Pyright type checking errors (0 errors, 0 warnings)
- Add proper Union types for GitHub API responses
- Fix OAuth attribute access with null-safety checks
- Add comprehensive type annotations throughout codebase

## Testing Improvements
- Expand test suite from 70 to 79 comprehensive tests
- Optimize test performance from minutes to ~1.5 seconds execution
- Add autouse fixtures for configuration mocking
- Fix slow callback route tests with proper OAuth mocking
- Update all test files to work with new configuration system

## Lint & Code Formatting
- Apply autopep8 and isort for consistent code formatting
- Remove unused imports and variables (F401, F841 errors)
- Fix import order and organization
- Resolve trailing whitespace and formatting issues
- Maintain 79-character line length where possible

## Documentation Updates
- Update CLAUDE.md with new architecture and configuration details
- Enhance TESTING.md with performance optimizations and new test info
- Document environment variable requirements and setup procedures
- Add comprehensive examples for all usage scenarios

+++ client_with_oauth.py
@@ -30,7 +30,8 @@ def setup_logging(verbose: bool = False):
         # Create handler for stdout output when verbose mode is enabled
         handler = logging.StreamHandler()
         handler.setLevel(logging.DEBUG)
-        formatter = logging.Formatter('%(filename)s:%(lineno)d - %(levelname)s - %(message)s')
+        formatter = logging.Formatter(
+            '%(filename)s:%(lineno)d - %(levelname)s - %(message)s')
         handler.setFormatter(formatter)
         logger.addHandler(handler)
 
@@ -114,7 +115,7 @@ async def get_device_code() -> dict[str, Any] | None:
             return None
 
 
-async def poll_for_token(device_code: str, interval: int = 5) -> str | None:
+async def poll_for_token(device_code: str, interval: float = 5) -> str | None:
     """Poll GitHub for the access token."""
     if not GITHUB_CLIENT_ID:
         raise ValueError("GITHUB_CLIENT_ID environment variable not set")
@@ -168,7 +169,8 @@ async def authenticate() -> str | None:
         logger.debug(" Using existing access token")
         return _access_token
 
-    # Check for Personal Access Token from environment variable (automated auth)
+    # Check for Personal Access Token from environment variable (automated
+    # auth)
     if GITHUB_PERSONAL_ACCESS_TOKEN:
         logger.debug("Using Personal Access Token from environment: %s",
                      GITHUB_PERSONAL_ACCESS_TOKEN)
@@ -194,8 +196,9 @@ async def authenticate() -> str | None:
     print(f" 2. Enter code: {device_info['user_code']}")
     print(" 3. Waiting for authentication...")
     print(f" (Code expires in {device_info.get('expires_in', 900)} seconds)")
-    print("\nAlternatively, set GITHUB_PERSONAL_ACCESS_TOKEN environment variable "
-          "with a Personal Access Token for automated authentication.")
+    print(
+        "\nAlternatively, set GITHUB_PERSONAL_ACCESS_TOKEN environment variable "
+        "with a Personal Access Token for automated authentication.")
 
     # Poll for token
     _access_token = await poll_for_token(
